Use long for delay everywhere related to @Background

+++ AndroidAnnotations/androidannotations-core/androidannotations-api/src/main/java/org/androidannotations/annotations/Background.java
@@ -191,7 +191,7 @@ public @interface Background {
 	 *
 	 * @return the delay of the execution
 	 */
-	int delay() default 0;
+	long delay() default 0;
 
 	/**
 	 * Serial execution group.
